Moved all duplicate code from mobile and desktop questionaire to util or parent files
